requirements.txt
torch
torchvision
timm
pandas
numpy
scikit-learn
albumentations
opencv-python
pillow
pydicom
tqdm
matplotlib
seaborn
torchmetrics
einops
faiss-cpu
sentence-transformers
transformers
accelerate
mlflow
wandb
fastapi
uvicorn[standard]
python-multipart
requests